chore(result): extract result behavior to module
